+++ static_precompiler/compilers/__init__.py
@@ -5,3 +5,4 @@ from .scss import *  # noqa
 from .less import *  # noqa
 from .stylus import *  # noqa
 from .handlebars import *  # noqa
+from .livescript import *  # noqa